pyproject.toml
[tool.poetry]
name = "port2ser"
version = "0.1.0"
description = ""
authors = ["Chen Yang <yangchen@thinkmore.info>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial-asyncio = "^0.5"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-asyncio = "^0.15.1"

[tool.poetry.scripts]
ser2port = 'port2ser.console:ser2port'
port2ser = 'port2ser.console:port2ser'


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
